backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
Django==4.2
django-crontab==0.7.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
gunicorn==21.2.0
idna==3.6
jedi==0.19.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
python-dotenv==1.0.1
